requirements.txt
# Minimal deps (keep it light)
requests>=2.31.0
PyYAML>=6.0.1
